+++ src/common.c
@@ -398,7 +398,7 @@ int Q_atoi (char *str)
 
 float Q_atof (char *str)
 {
-    double			val;
+    float			val;
     int             sign;
     int             c;
     int             decimal, total;
