+++ src/net_wins.c
@@ -61,7 +61,7 @@ WSADATA		winsockdata;
 
 //=============================================================================
 
-static double	blocktime;
+static float	blocktime;
 
 BOOL PASCAL FAR BlockingHook(void)  
 {
